Create api-types package to share API types
Created package based on API types to avoid having to copy types over to
each service that uses them. Updated the Dockerfiles for the bot and UI
services to account for the package which needs to be generated before
running `npm i`.

Updated the backend to use the types used to generate the new package
instead of the ones that were previously in the `shared` directory.
Removed types from `shared` directory.