(maint) Fix reactivity issue caused by error
When an error is caused by the fetcher function passed to
`createResource` the value of the error is also thrown when you try to
access the data value. In protected route, the first match tries to
access the `user()` value, but there is a chance that it could throw an
error (since the user may not be authenticated and the API returned a
401). Updated the component to first check if there is an error before
trying to access the value from createResource.